optimization of function to check if cycle is present in list

+++ 0x00-python-hello_world/10-check_cycle.c
@@ -6,7 +6,7 @@
 */
 int check_cycle(listint_t *list)
 {
-	int *p = list;
+	long int *p = list;
 
 	if (list == NULL)
 		return (0);
